Fix ListMixinCreator's prototype injection
