+++ client/models/mixins/ListingMixin.js
@@ -1,0 +1,38 @@
+import paginate from 'simple-pagination';
+
+
+const ListingMixin = {
+
+  _initializeListingMixin() {
+    this._listObjects = [];
+  },
+
+  getListObjects() {
+    return this._listObjects;
+  },
+
+  /*
+   * @param {number} rowsPerPage
+   * @param {number} specifiedPage - To be started from 1
+   * @return {object}
+   */
+  getListPagination(rowsPerPage, specifiedPage) {
+    const paginated = paginate(this._listObjects.length, rowsPerPage, specifiedPage);
+    const fromIndex = paginated.fromCount - 1;
+    const objects = this._listObjects.slice(fromIndex, fromIndex + rowsPerPage);
+    const indexedObjects = objects.map((object, i) => {
+      const serialNumber = paginated.fromCount + i;
+      return {
+        index: serialNumber - 1,
+        serialNumber,
+        object,
+      };
+    });
+    return Object.assign(paginated, {
+      objects,
+      indexedObjects,
+    });
+  },
+};
+
+export default ListingMixin;
